SimSpread

Documentation for SimSpread.

Base.split

SimSpread.AuPRC

SimSpread.AuROC

SimSpread.BEDROC

SimSpread._namedmatrix2matrix

SimSpread._parse_matrix

SimSpread.clamp

SimSpread.clean!

SimSpread.cutoff

SimSpread.cutoff

SimSpread.cutoff!

SimSpread.cutoff!

SimSpread.featurize

SimSpread.k

SimSpread.maxperformance

SimSpread.meanperformance

SimSpread.meanstdperformance

SimSpread.predict

SimSpread.predict

SimSpread.prepare

SimSpread.prepare!

SimSpread.read_namedmatrix

SimSpread.save

SimSpread.save

SimSpread.spread

Base.split — Method

split(DT::NamedArray, k::Int64, rng::Int64)

Split all possible D into k groups for cross-validation.

Long description

Split drugs D into k groups, extract their edges and append to cross-validation group.

Arguments

DT::AbstractMatrix: Drug-Target rectangular adjacency matrix.

k::Int64: Number of groups to use in data splitting.

rng::Int64: Seed used for data splitting.

SimSpread.AuPRC — Method

AuPRC(y::AbstractArray{Bool}, yhat::AbstractArray{Number})

Area under the Precision-Recall curve using the trapezoidal rule.

Arguments

y::AbstractArray: Binary class labels. 1 for positive class, 0 otherwise.

̂yhat::AbstractArray: Prediction values.

SimSpread.AuROC — Method

AuROC(y::AbstractArray{Bool}, yhat::AbstractVector{Number})

Area under the Receiver Operator Characteristic curve using the trapezoidal rule.

Arguments

y::AbstractArray: Binary class labels. 1 for positive class, 0 otherwise.

̂yhat::AbstractArray: Prediction values.

SimSpread.BEDROC — Method

BEDROC(y, yhat; rev = true, α = 20.0)

The Boltzmann Enhanced Descrimination of the Receiver Operator Characteristic (BEDROC) score is a modification of the Receiver Operator Characteristic (ROC) score that allows for a factor of early recognition.

Score takes a value in interval [0, 1] indicating degree to which the predictive model employed detects (early) the positive class.

Arguments

y::AbstractArray: Binary class labels. 1 for positive class, 0 otherwise.

̂yhat::AbstractArray: Prediction values.

rev::Bool: True if high values of $yhat$ correlates to positive class (default = true).

α::AbstractFloat: Early recognition parameter (default = 20.0).

References

Truchon, J.-F., & Bayly, C. I. (2007). Evaluating Virtual Screening Methods: Good and

Bad Metrics for the “Early Recognition” Problem. Journal of Chemical Information and Modeling, 47(2), 488–508. https://doi.org/10.1021/ci600426e

SimSpread._namedmatrix2matrix — Method

_namedmatrix2matrix(M::NamedMatrix)

Convert NamedMatrix to Base.Array with row / column of names

SimSpread._parse_matrix — Function

_parse_matrix(M::AbstractMatrix, rows::Bool, cols::Bool; type::Type)

Convert matrix with row/column names to NamedMatrix (assumes names are in firt row/column).

Arguments

M::AbstractMatrix : Matrix to parse

rows::Bool : Matrix has row names (default = false)

cols::Bool : Matrix has column names (default = false)

type::Type : Type of matrix values (default = Any)

SimSpread.clamp — Method

clamp(val::Number, vmin::Number, vmax::Number)

Restrict a value to a given range.

Arguments

val::Number : Value to clamp

vmin::Number : Value floor

vmax::Number : Value roof

SimSpread.clean! — Method

clean!(R::NamedArray, A::NamedArray, DT::NamedArray)

Flag errors from cross-validation splitting in place.

Arguments

R::NamedArray: Predicted drug-target interactions adjacency matrix

A::NamedArray: de novo NBI initial resources adjacency matrix

DT::NamedArray: Ground-truth drug-target interactions adjacency matrix

SimSpread.cutoff! — Method

cutoff!(M::AbstractVecOrMat{T}, α::T, weighted::Bool=false) where {T<:AbstractFloat}

Transform M based in SimSpread's similarity cutoff function, overwriting M.

Arguments

M::T : Matrix or Vector to transform

α::T : Similarity threshold

weighted::Bool : Apply weighting function to outcome (default = false)

SimSpread.cutoff! — Method

cutoff!(x::T, α::T, weighted::Bool=false) where {T<:AbstractFloat}

Transform x based in SimSpread's similarity cutoff function, overwriting x.

Arguments

x::T : Value to transform

α::T : Similarity matrix

weighted::Bool : Apply weighting function to outcome (default = false)

SimSpread.cutoff — Method

cutoff(x::AbstractMatrix{T}, α::T, weighted::Bool=false) where {T<:AbstractFloat}

Transform M based in SimSpread's similarity cutoff function.

Arguments

M::T : Matrix or Vector to transform

α::T : Similarity threshold

weighted::Bool : Apply weighting function to outcome (default = false)

SimSpread.cutoff — Method

cutoff(x::T, α::T, weighted::Bool=false) where {T<:AbstractFloat}

Transform x based in SimSpread's similarity cutoff function.

Arguments

x::T : Value to transform

α::T : Similarity matrix

weighted::Bool : Apply weighting function to outcome (default = false)

SimSpread.featurize — Method

featurize(M::NamedArray, α::Float64, β::Float64, weighted::Bool)

Convert continuous feature into binary feature based in 2 cutoffs: (i) α for strong-ties and (ii) β for weak-ties. Weighted version of function weights binary features with it's real value.

Arguments

M::AbtractMatrix: Continuous feature matrix

α::AbstractFloat: Strong-ties cutoff

β::AbstractFloat: Weak-ties cutoff

weighted::Bool: Flag for feature weighting using real value

SimSpread.k — Method

k(G::AbstractMatrix)

Get node degrees from adjacency matrix

Arguments

M::AbstractMatrix : Matrix to parse

SimSpread.maxperformance — Method

maxperf(confusion::ROCNums, metric::Function)

Get maximum performance of a given metric over a set of confusion matrices.

SimSpread.meanperformance — Method

meanperf(confusion::ROCNums, metric::Function)

Get mean performance of a given metric over a set of confusion matrices.

SimSpread.meanstdperformance — Method

meanstdperf(confusion::ROCNums, metric::Function)

Get mean and standard deviation performance of a given metric over a set of confusion matrices.

SimSpread.predict — Method

predict(A::NamedMatrix, B::NamedMatrix, names::Tuple; GPU::Bool)

TODO: Add short description to predict

Arguments

A::NamedMatrix: Compound-Feature-Drug-Target initial resources adjacency matrix

B::NamedMatrix: Feature-Drug-Target initial resources adjacency matrix

names::Tuple: Rows & columns named indices

GPU::Bool: (default = false)

SimSpread.predict — Method

predict(I::Tuple{T,T}, DT::T; GPU::Bool) where {T<:NamedArray}

TODO: Add short description to predict

Arguments

I::Tuple{NamedMatrix,NamedMatrix}: De novo initial resources adjacency matrices

DT::NamedMatrix: Drug-target adjacency matrix

GPU::Bool: Use GPU acceleration for calculation (default = false)

SimSpread.prepare! — Method

prepare!(DT::T, DF::T, Cs::AbstractVector) where {T<:NamedMatrix}

Prepare compound-feature-drug-target network adjacency matrix for de novo NBI prediction.

Arguments

DT::NamedMatrix: Drug-Target adjacency matrix

DF::NamedMatrix: Drug-Feature adjacency matrix

Cs::AbstractVector: Compounds to predict targets

Extended help

This implementation is for k-fold or leave-one-out cross-validation.

SimSpread.prepare — Method

prepare(dts::T, dfs::T) where {T<:Tuple{NamedMatrix,NamedMatrix}}

Prepare compound-feature-drug-target network adjacency matrix for de novo NBI predictions.

Arguments

dts::Tuple{NamedMatrix,NamedMatrix} : Drug-Target adjacency matrices

dfs::Tuple{NamedMatrix,NamedMatrix} : Drug-Feature adjacency matrices

Extended help

This implementation is for time-split cross-validation.

SimSpread.read_namedmatrix — Function

read_namedmatrix(filepath::String, valuetype::Type = FLoat64filepath::String, valuetype::Type)

Load a matrix with named indices as a NamedArray.

Arguments

filepath::String : File path of matrix to load

delimiter::Char : Delimiter character between values in matrix (default = ' ')

valuetype::Type : Type of values contained in matrix (default = Float64)

SimSpread.save — Method

save(filepath::String, fidx::Int64, C::AbstractVector, R::NamedMatrix, DT::NamedMatrix)

Store cross-valudation predictions as a table in the given file path.

Arguments

filepath::String : Output file path

fidx::Int64 : Numeric fold ID

C::AbstractVector : Test set compounds / ligands

R::NamedMatrix : Predicted drug-targetinteraction adjacency matrix

DT::NamedMatrix : Ground-truth drug-target interactions adjacency matrix

Extended help

Table format is:

Fold ID, Compound ID, Target ID, SimSpread Score, True-Positive state

SimSpread.save — Method

save(filepath::String, R::NamedMatrix, DT::NamedMatrix)

Store predictions as a table in the given file path.

Arguments

filepath::String: Output file path

R::NamedArray: Drug-target predictions matrix

DT::NamedMatrix: Drug-target interactions adjacency matrix

Extended help

Table format is:

fold, compound ID, target ID, score, TP

SimSpread.spread — Method

spread(F₀::AbstractMatrix{Float64})

Arguments

F₀::AbstractMatrix{Float64}: Trilayered SimSpread network adjacency matrix

References

Wu, et al (2016). SDTNBI: an integrated network and chemoinformatics tool for systematic prediction of drug–target interactions and drug repositioning. Briefings in Bioinformatics, bbw012. https://doi.org/10.1093/bib/bbw012

Vigil-Vásquez & Schüller (2022). De Novo Prediction of Drug Targets and Candidates by Chemical Similarity-Guided Network-Based Inference. International Journal of Molecular Sciences, 23(17), 9666. https://doi.org/10.3390/ijms23179666